W4C2
Used forms and javascript. Used the GET method to have javascript parse
the HTML to get the important elements from the URL to put into a table
on the next page. Added an id to the phone number group to be able to
hide it in the console as per class website instructions.